requirements.txt
fastapi==0.102.0
uvicorn==0.22.0
mysql-connector-python==8.0.29
bcrypt==4.0.1
pydantic==1.10.2
python-dotenv==0.21.0
